parts: guvcview: Fix broken L10N
Guvcview uses PACKAGE_LOCALE_DIR to bind the localization data dir, but
the value is defined as "$(prefix)/$(DATADIRNAME)/locale" instead of
"$(datadir)"/locale and thus guvcview search the data in /share/locale
instead of $SNAP/share/locale.  This patch sets the prefix instead of
the conservative datarootdir to workaround this issue.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>